Move paymentmethod package into the components package
COSDK-500
