+++ core/src/main/java/com/adyen/checkout/core/components/paymentmethod/MBWayPaymentMethod.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 12/5/2025.
  */
 
-package com.adyen.checkout.core.paymentmethod
+package com.adyen.checkout.core.components.paymentmethod
 
 import com.adyen.checkout.core.PaymentMethodTypes
 import com.adyen.checkout.core.exception.ModelSerializationException
