+++ core/src/main/java/com/adyen/checkout/core/components/paymentmethod/PaymentComponentState.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 16/5/2025.
  */
 
-package com.adyen.checkout.core.paymentmethod
+package com.adyen.checkout.core.components.paymentmethod
 
 import com.adyen.checkout.core.data.PaymentComponentData
