+++ core/src/main/java/com/adyen/checkout/core/data/PaymentComponentData.kt
@@ -8,6 +8,7 @@
 
 package com.adyen.checkout.core.data
 
+import com.adyen.checkout.core.components.paymentmethod.PaymentMethodDetails
 import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.internal.data.model.ModelObject
@@ -15,7 +16,6 @@ import com.adyen.checkout.core.internal.data.model.ModelUtils.deserializeOpt
 import com.adyen.checkout.core.internal.data.model.ModelUtils.serializeOpt
 import com.adyen.checkout.core.internal.data.model.getBooleanOrNull
 import com.adyen.checkout.core.internal.data.model.getStringOrNull
-import com.adyen.checkout.core.paymentmethod.PaymentMethodDetails
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
