+++ example-app/src/main/java/com/adyen/checkout/example/ui/v6/V6ViewModel.kt
@@ -14,7 +14,7 @@ import com.adyen.checkout.core.CheckoutCallback
 import com.adyen.checkout.core.CheckoutConfiguration
 import com.adyen.checkout.core.CheckoutResult
 import com.adyen.checkout.core.Environment
-import com.adyen.checkout.core.paymentmethod.PaymentComponentState
+import com.adyen.checkout.core.components.paymentmethod.PaymentComponentState
 import com.adyen.checkout.example.BuildConfig
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
